rename server to createServer
